add transaction type and parse transfer tx

+++ ledger/nuls-ledger/src/main/java/io/nuls/ledger/model/FreezeLockTimeState.java
@@ -1,8 +1,5 @@
 package io.nuls.ledger.model;
 
-import io.nuls.ledger.utils.ByteUtil;
-import io.nuls.ledger.utils.RLP;
-import io.nuls.ledger.utils.RLPList;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -28,7 +25,7 @@ public class FreezeLockTimeState {
      */
     @Setter
     @Getter
-    private long amount;
+    private BigInteger amount;
 
     /**
      * 锁定时间
